+++ scripts/docs.sh
@@ -28,3 +28,4 @@ mkdir -p docs/moonblade
 $XAN help cheatsheet --md > docs/moonblade/cheatsheet.md
 $XAN help functions --md > docs/moonblade/functions.md
 $XAN help aggs --md > docs/moonblade/aggs.md
+$XAN help scraping --md > docs/moonblade/scraping.md